test(core): adapted funding tests

+++ src/UsdnProtocol/UsdnProtocolCore.sol
@@ -17,6 +17,8 @@ import {
 import { SignedMath } from "src/libraries/SignedMath.sol";
 import { DoubleEndedQueue } from "src/libraries/DoubleEndedQueue.sol";
 
+import { console2 } from "forge-std/Test.sol";
+
 abstract contract UsdnProtocolCore is IUsdnProtocolCore, UsdnProtocolStorage {
     using SafeERC20 for IERC20Metadata;
     using SafeCast for uint256;
@@ -211,11 +213,11 @@ abstract contract UsdnProtocolCore is IUsdnProtocolCore, UsdnProtocolStorage {
         if (oldLongExpo_ <= 0) {
             // if oldLongExpo is negative, then we cap the imbalance index to -1
             // oldVaultExpo is always positive
-            return (-int256(_fundingSF) * int256(10 ** FUNDING_RATE_DECIMALS - FUNDING_SF_DECIMALS) + ema, oldLongExpo_);
+            return (-int256(_fundingSF * 10 ** (FUNDING_RATE_DECIMALS - FUNDING_SF_DECIMALS)) + ema, oldLongExpo_);
         } else if (oldVaultExpo == 0) {
             // if oldVaultExpo is zero (can't be negative), then we cap the imbalance index to 1
             // oldLongExpo must be positive in this case
-            return (int256(_fundingSF) * int256(10 ** FUNDING_RATE_DECIMALS - FUNDING_SF_DECIMALS) + ema, oldLongExpo_);
+            return (int256(_fundingSF * 10 ** (FUNDING_RATE_DECIMALS - FUNDING_SF_DECIMALS)) + ema, oldLongExpo_);
         }
 
         // starting here, oldLongExpo and oldVaultExpo are always strictly positive
